pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-project-initializer"
version = "0.1.0"
description = "Local-first CLI bootstrap tool for PRD-driven project initialization."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jsonschema",
  "pyyaml"
]

[project.scripts]
initializer = "initializer.cli:main"

[tool.setuptools.packages.find]
include = ["initializer*"]
